requirements.txt
beautifulsoup4
requests
pypdf
scikit-learn
faiss-cpu
azure-cosmos
azure-identity
azure-search-documents
openai
python-dotenv
aiohttp
numpy
scipy
tqdm
voila
ipywidgets
plotly
pandas
